MpaDecGMpaDecFilter: PS2ADPCM декодер теперь полностью на целочисленной арифметике.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@5267 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/transform/MpaDecFilter/MpaDecFilter.h
@@ -56,7 +56,7 @@ enum {
 
 struct ps2_state_t {
 	bool sync;
-	double a[2], b[2];
+	int a[2], b[2];
 	ps2_state_t() {
 		reset();
 	}
